move get_all_accounts to chain_plugin, sync api, add pagination, restore account_query_db
