+++ plugins/chain_api_plugin/chain_api_plugin.cpp
@@ -140,15 +140,15 @@ void chain_api_plugin::plugin_startup() {
       CHAIN_RW_CALL_ASYNC(push_transaction, chain_apis::read_write::push_transaction_results, 202, http_params_types::params_required),
       CHAIN_RW_CALL_ASYNC(push_transactions, chain_apis::read_write::push_transactions_results, 202, http_params_types::params_required),
       CHAIN_RW_CALL_ASYNC_V2(send_transaction, chain_apis::read_write::push_transaction_results, 202, http_params_types::params_required),
-      CHAIN_RW_CALL_ASYNC(send_transaction, chain_apis::read_write::send_transaction_results, 202, http_params_types::params_required)
+      CHAIN_RW_CALL_ASYNC(send_transaction, chain_apis::read_write::send_transaction_results, 202, http_params_types::params_required),
+      CHAIN_RO_CALL_WITH_400(get_all_accounts, 200, http_params_types::params_required)
    });
 
    if (chain.account_queries_enabled()) {
       _http_plugin.add_async_api({
          CHAIN_RO_CALL_WITH_400(get_accounts_by_authorizers, 200, http_params_types::params_required),
-         CHAIN_RO_CALL(get_all_accounts, 200, http_params_types::no_params_required)
       });
-   }
+   }   
 }
 
 void chain_api_plugin::plugin_shutdown() {}
